+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+- Improve negotiation protocol to allow maker to identify a taken order by swap id.
+
 ## [0.7.0] - 2019-11-27
 
 ### Changed